src/main/java/com/customerservice/CustomerService/service/CustomerService.java
package com.customerservice.CustomerService.service;

import com.customerservice.CustomerService.model.Customer;
import com.customerservice.CustomerService.repository.CustomerRepository;
import org.springframework.stereotype.Service;
import java.util.List;

@Service
public class CustomerService {

    private CustomerRepository customerRepository;

    public CustomerService(CustomerRepository customerRepository) {
        this.customerRepository = customerRepository;
    }

    public List<Customer> getCustomers() {
        return customerRepository.findAll();
    }

    public Customer addCustomer(Customer customer) {
        return customerRepository.save(customer);
    }

}
